[NCL-4810] Add support for custom parameters for Gradle builds
This reuses the `CUSTOM_PME_PARAMETERS` variable and is subject
for refactoring after 1.7 release.
