+++ repour/adjust/pme_provider.py
@@ -92,37 +92,6 @@ def get_pme_provider(execution_name, pme_jar_path, pme_parameters, output_to_log
                 return None, None
 
     @asyncio.coroutine
-    def get_extra_parameters(extra_adjust_parameters):
-        """
-        Get the extra PME parameters from PNC
-        If the PME parameters contain '--file=<folder>/pom.xml', then extract that folder
-        and remove that --file option from the list of extra params.
-        In PME 2.11 and PME 2.12, there is a bug where that option causes the file target/pom-manip-ext-result.json
-        to be badly generated. Fixed in PME 2.13+
-        See: PRODTASKS-361
-        Returns: tuple<list<string>, string>: list of params(minus the --file option), and folder where to run PME
-        If '--file' option not used, the folder will be an empty string
-        """
-        subfolder = ''
-
-        paramsString = extra_adjust_parameters.get("CUSTOM_PME_PARAMETERS", None)
-        if paramsString is None:
-            return [], subfolder
-        else:
-            params = shlex.split(paramsString)
-            for p in params:
-                if p[0] != "-":
-                    desc = ('Parameters that do not start with dash "-" are not allowed. '
-                            + 'Found "{p}" in "{params}".'.format(**locals()))
-                    raise exception.AdjustCommandError(desc, [], 10, stderr=desc)
-                if p.startswith("--file"):
-                    subfolder = p.replace("--file=", "").replace("pom.xml", "")
-
-            params_without_file_option = [p for p in params if not p.startswith("--file=")]
-
-            return params_without_file_option, subfolder
-
-    @asyncio.coroutine
     def adjust(repo_dir, extra_adjust_parameters, adjust_result):
         nonlocal execution_name
 
@@ -134,7 +103,7 @@ def get_pme_provider(execution_name, pme_jar_path, pme_parameters, output_to_log
         if specific_indy_group:
             temp_build_parameters.append("-DrestRepositoryGroup=" + specific_indy_group)
 
-        extra_parameters, subfolder = yield from get_extra_parameters(extra_adjust_parameters)
+        extra_parameters, subfolder = yield from util.get_extra_parameters(extra_adjust_parameters)
 
         # readjust the repo_dir to run PME from the folder where the root pom.xml is located
         # See: PRODTASKS-361
